requirements.txt
accelerate==0.16.0
anyio==3.6.2
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
diffusers==0.13.1
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.92.0
filelock==3.9.0
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.12.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
nest-asyncio==1.5.6
numpy==1.24.2
orjson==3.8.6
packaging==23.0
Pillow==9.4.0
psutil==5.9.4
pydantic==1.10.5
pyngrok==5.2.1
python-dotenv==0.21.1
python-multipart==0.0.5
PyYAML==6.0
redis==4.5.1
regex==2022.10.31
requests==2.28.2
rfc3986==1.5.0
rq==1.13.0
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
tokenizers==0.13.2
torch==1.13.1
tqdm==4.64.1
transformers==4.26.1
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
zipp==3.14.0
